feat(BA-7178): restructure into devices and kernel_devices tables
