+++ changes/13464.feature.md
@@ -1,1 +1,1 @@
-Add the kernel_attached_devices table recording devices attached to each kernel as typed rows
+Add the devices and kernel_devices tables recording devices discovered on agents and their per-kernel attachments as typed rows
